Moving tests to better names, prepping for some mild speedups for lookup.
